Added navigation, multiple svg, fixed various things

+++ app/src/main/java/com/example/pagekeeper/pages/domain/library/XmlParser.kt
@@ -6,6 +6,6 @@ import com.example.pagekeeper.core.domain.util.Result
 
 interface XmlParser {
     suspend fun parseBook(uri: Uri): Result<Book, ParserError>
-    suspend fun parseBookBodyFile(bookId: Int): Result<Unit, ParserError>
+    suspend fun parseBookBodyFile(bookId: Long): Result<Unit, ParserError>
     suspend fun deleteBook(path: String): Boolean?
 }
