+++ app/src/main/java/com/example/pagekeeper/pages/presentation/library/LibraryEvent.kt
@@ -6,5 +6,5 @@ sealed interface LibraryEvent {
     data object OnImportBook: LibraryEvent
     data class OnShareBook(val path: String): LibraryEvent
     data class OnShareMultipleBooks(val paths: List<String>): LibraryEvent
-    data class OnBookSelected(val bookId: Int): LibraryEvent
+    data class OnBookSelected(val bookId: Long): LibraryEvent
 }
